Updates stubs for how python3.12 with pybind11-stubgen makes them
