+++ python/stubs/ACEtk/ACEtk/photoatomic.pyi
